requirements.txt
fastapi
pymongo
pydantic
datetime
pyfcm
uvicorn
gunicorn
jinja2
python-multipart
pyjwt
passlib